US9718 Starting to Refactor CMS Service to es6
